added sweet alert to all the crud operations

+++ src/Pages/My Bookings/MyBookings.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Providers/AuthProvider";
 import BookingRow from "../../Components/BookingRow";
 // import BookingRow from "../../Components/BookingRow";
+import Swal from 'sweetalert2'
 
 const MyBookings = () => {
     const { user } = useContext(AuthContext)
@@ -15,21 +16,43 @@ const MyBookings = () => {
     }
         , [user?.email])
 
-    const handleDelete = id => {
-
-        fetch(`http://localhost:5000/bookings/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount > 0) {
-                    const remaining = bookings.filter(booking => booking._id !== id);
-                    setBookings(remaining);
+        const handleDelete = _id => {
+            console.log(_id);
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                console.log(result)
+                if (result.isConfirmed==true) {
+    
+    
+                    fetch(`http://localhost:5000/bookings/${_id}`, {
+                        method: 'DELETE'
+                    })
+                        .then(res => res.json())
+                        .then(data => {
+                            console.log(data);
+                            if (data.deletedCount > 0) {
+                                Swal.fire(
+                                    'Deleted!',
+                                    'Your Coffee has been deleted.',
+                                    'success'
+                                )
+                                const remaining = bookings.filter(booking => booking._id !== id);
+                                setBookings(remaining);
+                            }
+                        })
+    
                 }
             })
-
         }
+    
+
